Give wifly open call a longer timeout
After sending the open command to the wifly command prompt, the open()
function expected to see a prompt again immediately via a getPrompt()
call. This isn't realistic on a network; establishing a TCP connection
is still subject to the laws of physics. A one second timeout seems to
be working much better.