Remove wait for reboot when open() fails
The while(1) that used to be in open() simply hangs the Arduino sketch
on an infinite loop. If the WiFly card is rebooting, it's not bringing
the Arduino along for the ride, so it's pointless to wait for a reset
that's never coming. open() simply returns false now to indicate that it
was unable to open a connection.