Update to Pusher protocol version 6
The only change is to add an explicit protocol=6 URL argument when
connecting to Pusher, so the Pusher system doesn't assume we're using a
deprecated version of the javascript library. The Arduino library
reports itself as the js library, which formerly would result in Pusher
inserting the protocol version for us. This is no longer the case, so
protocol is now stated up front during the connection.